terminado facturacion

+++ src/facturacion/ItemDesglose.java
@@ -5,16 +5,16 @@ import java.time.LocalDate;
 public class ItemDesglose {
 	
 	private String detalle;
-	private int monto;
+	private double monto;
 	private LocalDate fecha;
 	
-	public ItemDesglose(String detalle, int monto, LocalDate fecha) {
+	public ItemDesglose(String detalle, double monto, LocalDate fecha) {
 		this.detalle = detalle;
 		this.monto = monto; 
 		this.fecha = fecha;
 	}
 	
-	public int getMonto() {
+	public double getMonto() {
 		return monto;
 	}
